data/requirements.txt
pandas==2.3.2
mysql-connector-python==9.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.31.0 